codes about Bookmark refactoring it is abided by the principles of object-oriented
