+++ src/main/java/com/example/thehealingmeal/bookmark/service/BookmarkGenerater.java
@@ -7,21 +7,22 @@ import com.example.thehealingmeal.bookmark.repository.BookmarkRepository;
 import com.example.thehealingmeal.bookmark.repository.SnackBookmarkRepository;
 import com.example.thehealingmeal.member.domain.User;
 import com.example.thehealingmeal.member.repository.UserRepository;
-import com.example.thehealingmeal.menu.api.dto.MenuResponseDto;
-import com.example.thehealingmeal.menu.api.dto.SnackOrTeaResponseDto;
-import com.example.thehealingmeal.menu.domain.*;
-import com.example.thehealingmeal.menu.domain.repository.*;
+import com.example.thehealingmeal.menu.domain.MenuForUser;
+import com.example.thehealingmeal.menu.domain.SideDishForUserMenu;
+import com.example.thehealingmeal.menu.domain.SnackOrTea;
+import com.example.thehealingmeal.menu.domain.repository.MenuRepository;
+import com.example.thehealingmeal.menu.domain.repository.SideDishForUserMenuRepository;
+import com.example.thehealingmeal.menu.domain.repository.SnackOrTeaMenuRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class BookmarkService {
+public class BookmarkGenerater {
     private final UserRepository userRepository;
     private final MenuRepository menuRepository;
     private final BookmarkRepository bookmarkRepository;
@@ -74,38 +75,6 @@ public class BookmarkService {
         bookmarkRepository.save(bookmark);
     }
 
-    // 아점저 메뉴 즐겨찾기 조회
-    public List<MenuResponseDto> menuBookmarkList(Long userId) {
-        List<Bookmark> bookmarkList = bookmarkRepository.findByUserId(userId);
-
-        List<MenuResponseDto> menuResponseDtos = new ArrayList<>(); // 반환할 값
-
-        for (Bookmark bookmark : bookmarkList) {
-            MenuResponseDto menuResponseDto = MenuResponseDto.createMenu(
-                    bookmark.getId(),
-                    bookmark.getMain_dish(),
-                    bookmark.getImageURL(),
-                    bookmark.getRice(),
-                    bookmark.getMeals(),
-                    bookmark.getSideDishForUserMenu(),
-                    bookmark.getKcal(),
-                    bookmark.getProtein(),
-                    bookmark.getCarbohydrate(),
-                    bookmark.getFat()
-            );
-            menuResponseDtos.add(menuResponseDto);
-        }
-
-        return menuResponseDtos;
-    }
-
-    // 아점저 메뉴 즐겨찾기 삭제.
-    @Transactional
-    public void deleteMenuBookmark(Long bookmarkId) {
-        Bookmark bookmark = bookmarkRepository.findById(bookmarkId).orElseThrow();
-        bookmarkRepository.delete(bookmark);
-    }
-
      //간식 메뉴 즐겨찾기 저장.
     @Transactional
     public void createSnackBookmark(Long userId, BookmarkRequestDto requestDto) {
@@ -135,34 +104,5 @@ public class BookmarkService {
                 .build();
 
         snackBookmarkRepository.save(snackBookmark);
-    }
-//
-//    // 간식 메뉴 즐겨찾기 조회
-    public List<SnackOrTeaResponseDto> snackBookmarkList(Long userId) {
-        List<SnackBookmark> snackBookmarkList = snackBookmarkRepository.findByUserId(userId);
-
-        List<SnackOrTeaResponseDto> snackOrTeaResponseDtos = new ArrayList<>(); // 반환할 값
-
-
-        for (SnackBookmark snackBookmark : snackBookmarkList) {
-            SnackOrTeaResponseDto snackOrTeaResponseDto = SnackOrTeaResponseDto.createMenu(
-                    snackBookmark.getId(),
-                    snackBookmark.getSnack_or_tea(),
-                    snackBookmark.getImageUrl(),
-                    snackBookmark.getMeals(),
-                    snackBookmark.getKcal(),
-                    snackBookmark.getProtein(),
-                    snackBookmark.getCarbohydrate(),
-                    snackBookmark.getFat()
-            );
-            snackOrTeaResponseDtos.add(snackOrTeaResponseDto);
-        }
-        return snackOrTeaResponseDtos;
-    }
-    // 간식 메뉴 즐겨찾기 삭제.
-    @Transactional
-    public void deleteSnackBookmark(Long snackBookmarkId) {
-        SnackBookmark snackBookmark = snackBookmarkRepository.findById(snackBookmarkId).orElseThrow();
-        snackBookmarkRepository.delete(snackBookmark);
     }
 }
